separando swagger del index
